make Deflate always available.
