+++ Project.toml
@@ -3,6 +3,7 @@ uuid = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 version = "0.5.15"
 
 [deps]
+CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 Mmap = "a63ad114-7e13-5084-954f-fe012c677804"
@@ -11,34 +12,32 @@ PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 ScopedValues = "7e506255-f358-4e82-b7e4-beb19740aa63"
 
 [weakdeps]
-UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
-CodecBzip2 = "523fee87-0ab8-5b00-afb7-3ecf72e48cfd"
-CodecZstd = "6b39b394-51ab-5f42-8807-6242bab2b4c2"
-CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
-CodecLz4 = "5ba52731-8f18-5e0d-9241-30f10d1ec561"
-bitshuffle_jll = "228fe19c-1b83-5282-a626-13744502a320"
 Blosc = "a74b3585-a348-5f62-a45c-50e91977d574"
+CodecBzip2 = "523fee87-0ab8-5b00-afb7-3ecf72e48cfd"
+CodecLz4 = "5ba52731-8f18-5e0d-9241-30f10d1ec561"
+CodecZstd = "6b39b394-51ab-5f42-8807-6242bab2b4c2"
+UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
+bitshuffle_jll = "228fe19c-1b83-5282-a626-13744502a320"
 
 [extensions]
-UnPackExt = "UnPack"
-CodecBzip2Ext = "CodecBzip2"
-CodecZstdExt = "CodecZstd"
-CodecZlibExt = "CodecZlib"
-CodecLz4Ext = "CodecLz4"
-bitshuffle_jllExt = "bitshuffle_jll"
 BloscExt = "Blosc"
+CodecBzip2Ext = "CodecBzip2"
+CodecLz4Ext = "CodecLz4"
+CodecZstdExt = "CodecZstd"
+UnPackExt = "UnPack"
+bitshuffle_jllExt = "bitshuffle_jll"
 
 [compat]
-CodecZstd = "0.8"
-CodecZlib = "0.7.8"
 Blosc = "0.7.3"
-bitshuffle_jll = "0.5.2"
-CodecLz4 = "0.4.6"
 CodecBzip2 = "0.8.5"
+CodecLz4 = "0.4.6"
+CodecZlib = "0.7.8"
+CodecZstd = "0.8"
 FileIO = "1.5"
 MacroTools = "0.5.10"
 OrderedCollections = "1"
 PrecompileTools = "1"
-UnPack = "1"
 ScopedValues = "1.3"
+UnPack = "1"
+bitshuffle_jll = "0.5.2"
 julia = "1.9"
