--- filterpkgs/JLD2Deflate/Project.toml
@@ -1,12 +1,0 @@
-name = "JLD2Deflate"
-uuid = "ccee2bf6-93a5-4c76-adcc-6b5f066a3aa3"
-authors = ["Jonas Isensee <jonas.isensee@web.de>"]
-version = "0.1.0"
-
-[deps]
-CodecZlib = "944b1d66-785c-5afd-91f1-9de20f533193"
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
-
-[compat]
-CodecZlib = "0.7.8"
-JLD2 = "0.5.15"
