--- filterpkgs/JLD2Deflate/README.md
@@ -1,5 +1,0 @@
-# JLD2Deflate.jl
-
-Implements the default deflate (compression) filter for [JLD2.jl](https://github.com/JuliaIO/JLD2.jl).
-
-For more information see e.g. [Using HDF5 Filters](https://hdfgroup.github.io/hdf5/develop/_h5_d__u_g.html#title6).
