+++ src/JLD2.jl
@@ -7,7 +7,7 @@ export load, save
 using ScopedValues: ScopedValue, with
 using PrecompileTools: @setup_workload, @compile_workload
 export jldopen, @load, @save, save_object, load_object, jldsave
-export Shuffle
+export Shuffle, Deflate
 
 include("types.jl")
 include("macros_utils.jl")
@@ -495,7 +495,7 @@ include("datatypes.jl")
 include("datalayouts.jl")
 
 include("Filters.jl")
-using .Filters: WrittenFilterPipeline, FilterPipeline, iscompressed, Shuffle
+using .Filters: WrittenFilterPipeline, FilterPipeline, iscompressed, Shuffle, Deflate
 
 include("datasets.jl")
 include("global_heaps.jl")
